+++ src/core/YUVData.cpp
@@ -51,11 +51,11 @@ class RasterYUVData : public YUVData {
     return data.size();
   }
 
-  const void* getBaseAddressAt(int planeIndex) const override {
+  const void* getBaseAddressAt(size_t planeIndex) const override {
     return data[planeIndex];
   }
 
-  size_t getRowBytesAt(int planeIndex) const override {
+  size_t getRowBytesAt(size_t planeIndex) const override {
     return rowBytes[planeIndex];
   }
